Word ranking, big fixes
Lots of fixes in this commit. Overhauled the next input picker system to prioritize words that contain more common letters that haven't been used yet, which provides better suggestions than the old ranking system. Also added a fix for if the user inputs bad information and it results in no words being a match.